Refactored SignUp functionality to provide error when signing up. ViewModel classes names refactored.

+++ app/src/main/java/project/lonelywheeler/db/entity/offer/OfferBasicInfoEntity.kt
@@ -6,7 +6,7 @@ import java.util.*
 class OfferBasicInfoEntity
 constructor(
     val title: String,
-    val value: Double,
+    val value: Int,
     val dateModified: Long,
     val model: String?,
     val brand: String,
